cdk-llm-observability-ref-impl-rag/container/app/requirements.txt
requests
streamlit
langchain
langchain-community
langchain-aws
requests-aws4auth
opensearch-py
boto3
pandas
textstat
spacy
traceloop-sdk
nltk
pydantic
tiktoken
evaluate
transformers
mlflow
sagemaker-mlflow
